Drop generated/ and use libi3ds_asn1
This drops the files from generated and instead ads the headers and
pre-compiled .so to the mix. It also sets the namespace to reduce the
fallout of this change.

Default behaviour is to look in the standard folders for a pre-isntalled
libi3ds_asn1, but it is also possible to instrcuct cmake to look in a
specific folder for the ASN.1 (en|de)codder.