update petamake to handle 2019.x sysroot
The sysroot has changed slightly and the environment we source will
populate variables used by cmake to set CMAKE_FIND_ROOT_PATH. When we
export this from petamake, we end up with a double sysroot, causing
build to fail.

We solve this by bypassing this and setting BUILD_SYSROOT instead, and
cmake will use this over FIND_ROOT_PATH whenever available.

This has been tested with both 2019.1 and 2019.2