Add simple camera-forwarder
Some nodes (pointcloud in particular) will take a camera-stream and
compute a pointcloud. When not operating as pointcloud-node, it will be
useful if it can replublish the images with a normal flash.

For debugging, it is also useful if it can be set to publish the images
it uses to compute the pointcloud.